CHANGE: network creating moved to GameStartActivity.java
