+++ app/src/main/AndroidManifest.xml
@@ -1,8 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.mia.phase10">
- <uses-permission android:name="android.permission.ACCESS_WIFI_STATE"></uses-permission>
-   <uses-permission android:name="android.permission.INTERNET"></uses-permission>
+
+    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE"></uses-permission>
+    <uses-permission android:name="android.permission.INTERNET"></uses-permission>
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -10,9 +12,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".GameActivity"
-            android:screenOrientation="landscape"
-            android:parentActivityName=".MainActivity">
+        <activity
+            android:name=".GameActivity"
+            android:parentActivityName=".MainActivity"
+            android:screenOrientation="landscape">
             <!-- The meta-data tag is required if you support API level 15 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
@@ -21,6 +24,9 @@
         <activity
             android:name=".ShufflingActivity"
             android:screenOrientation="landscape" />
+        <activity
+            android:name=".GameStartActivity"
+            android:screenOrientation="portrait" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
